backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
bcrypt==4.1.1
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
requests==2.31.0